Updated the test project to reflect the structural changes made in the infrastructure.
