+++ Listen2MeRefined.Tests/Media/MusicPlayer/PlaybackQueueServiceTests.cs
@@ -1,7 +1,7 @@
 using Listen2MeRefined.Infrastructure.Data.Models;
 using Listen2MeRefined.Infrastructure.Media.MusicPlayer;
 
-namespace Listen2MeRefined.Tests.Media;
+namespace Listen2MeRefined.Tests.Media.MusicPlayer;
 
 public class PlaybackQueueServiceTests
 {
